move use effect to main-content.tsx
